Move dispatcher.py to dispatcher_haigha.py
