+++ amqpdispatcher/dispatcher_haigha.py
